+++ .github/workflows/release.yaml
@@ -1,0 +1,23 @@
+name: release
+
+on:
+  pull_request:
+    branches:
+      - master
+    types: ["closed", "synchronize", "opened"]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+
+  test:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v4
+      - name: Install dependencies
+        run: pnpm install
+      - name: Run tests
+        run: pnpm test
