--- packages/config/src/index.ts
@@ -1,1 +1,0 @@
-export * from "./vite.base.config.js"
