+++ packages/core/package.json
@@ -17,7 +17,6 @@
   "description": "",
   "devDependencies": {
     "@halvaradop/auth-init": "^0.0.3",
-    "@halvaradop/repo-utils": "workspace:*",
     "@halvaradop/repo-config": "workspace:*"
   },
   "files": [
