--- packages/core/testing/index.test.js
@@ -1,4 +1,0 @@
-import { test, expect } from "vitest"
-test("core package", () => {
-    expect(0).toEqual(0)
-})
