--- packages/core/testing/index.test.ts
@@ -1,5 +1,0 @@
-import { test, expect } from "vitest"
-
-test("core package", () => {
-    expect(0).toEqual(0)
-})
