+++ packages/core/vite.config.ts
@@ -1,0 +1,4 @@
+import { defineConfig } from "vite"
+import { config } from "@halvaradop/repo-config/vite.config.base"
+
+export default defineConfig(config)
