--- packages/css/vite.config.js
@@ -1,6 +1,0 @@
-import { defineConfig } from "vite"
-import { viteDefaultConfig } from "@halvaradop/repo-config"
-
-export default defineConfig({
-    ...viteDefaultConfig,
-})
