+++ packages/jsx/README.md
@@ -1,3 +1,4 @@
 # @halvaradop/repo-jsx
-**WARNING:** This repository contains experimental code that may or may not work as intended. 
+
+**WARNING:** This repository contains experimental code that may or may not work as intended.
 Proceed with caution (and at your own risk)!
