+++ packages/jsx/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@halvaradop/repo-jsx",
-  "version": "0.4.0",
+  "version": "0.6.0",
   "private": false,
   "description": "",
   "type": "module",
