+++ packages/jsx/tsconfig.json
@@ -2,7 +2,8 @@
   "extends": "@halvaradop/repo-config/tsconfig.base.json",
   "compilerOptions": {
     "outDir": "dist",
-    "jsx": "react-jsx"
+    "jsx": "react-jsx",
+    "composite": true
   },
   "include": ["src"],
   "exclude": ["node_modules", "dist"]
