--- packages/jsx/tsup.config.js
@@ -1,13 +1,0 @@
-import { defineConfig } from "tsup"
-
-const config = {
-    entry: ["src/index.tsx"],
-    clean: true,
-    dts: true,
-    outDir: "dist",
-    external: ["tailwindcss"],
-    format: ["esm", "cjs"],
-    minify: true,
-}
-
-export default defineConfig(config)
