+++ packages/jsx/tsup.config.ts
@@ -1,0 +1,4 @@
+import { defineConfig } from "tsup"
+import { config } from "@halvaradop/repo-config/tsup.config"
+
+export default defineConfig(config)
