--- packages/utils/package.json
@@ -1,52 +1,0 @@
-{
-  "name": "@halvaradop/repo-utils",
-  "version": "1.3.1",
-  "private": false,
-  "type": "module",
-  "main": "dist/index.d.ts",
-  "module": "dist/utils/utils.js",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.d.ts",
-      "require": "./dist/index.d.cts"
-    },
-    "./utils": {
-      "types": "./dist/utils/utils.d.ts",
-      "import": "./dist/utils/utils.js",
-      "require": "./dist/utils/utils.cjs"
-    },
-    "./package.json": "./package.json"
-  },
-  "scripts": {
-    "dev": "tsup --watch",
-    "test": "vitest",
-    "build:js": "tsup src/utils.ts --format cjs,esm --minify --dts --outDir dist/utils",
-    "build:test": "tsup test/**/*.ts --format cjs,esm --minify --dts --outDir dist/test",
-    "build:test:watch": "tsup test/**/*.ts --format cjs,esm --dts --outDir dist/test --watch",
-    "build": "tsup && pnpm build:js && pnpm build:test",
-    "build:vite": "vite build",
-    "build:watch": "tsup --watch",
-    "start": "node dist/index.js",
-    "config": "tsc --showConfig"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "description": "",
-  "devDependencies": {
-    "@halvaradop/repo-config": "workspace:*",
-    "@halvaradop/shortify-cli": "^0.0.3",
-    "postcss": "^8.4.47",
-    "tailwindcss": "^3.4.10",
-    "tsup": "^8.2.4"
-  },
-  "files": [
-    "dist",
-    "package.json",
-    "README.md"
-  ],
-  "engines": {
-    "node": ">=18.0.0"
-  }
-}
