--- packages/utils/src/declaration.ts
@@ -1,7 +1,0 @@
-export interface Params {
-    searchParams: URLSearchParams
-    url: string
-    query: string
-}
-
-export type PartialParams = Partial<Params>
