--- packages/utils/src/types.ts
@@ -1,3 +1,0 @@
-export type ArgsMethod = (...args: any) => void
-export type ArgsAsyncMethod = Promise<ArgsMethod>
-export type ReturnType<T extends ArgsMethod> = T extends (...args: any) => infer Return ? Return : never
