--- packages/utils/src/utils.ts
@@ -1,13 +1,0 @@
-/**
- *
- * @param external
- */
-export const externalAsync = async (external: (...args: any) => void) => {
-    // fetching data
-    external()
-}
-
-export const config: Record<string, string> = {
-    path: "./src",
-    name: "utils",
-}
