--- packages/utils/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "@halvaradop/repo-config/tsconfig.base.json",
-  "compilerOptions": {
-    "outDir": "dist",
-  },
-  "include": ["src/**/*", "test/**/*"],
-  "exclude": ["dist"]
-}
